feat: isolate restaurant profiles, menu items, categories, users, and stock logs by active restaurantId on shared devices
- Change RestaurantProfileEntity primary key id to hold actual restaurantId

- Add Room migration v45 to migrate existing profile id = restaurant_id

- Filter categories, menu items, variants, users, and stock logs by active restaurantId in all DAO queries

- Inject and supply active restaurantId from SessionManager in all repositories

- Verified via full compilation, unit tests run, and debug APK installation

+++ Android/app/src/main/java/com/khanabook/lite/pos/data/local/entity/RestaurantProfileEntity.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 @Entity(tableName = "restaurant_profile")
 data class RestaurantProfileEntity(
     @SerializedName("localId") @PrimaryKey
-    val id: Int = 1,
+    val id: Long = 1L,
     @ColumnInfo(name = "shop_name")
     val shopName: String? = null,
     @ColumnInfo(name = "shop_address")
